Fix incompatibility of actor-critic LSTM methods
Summary:
This diff fixes the incompatibility of actor-critic methods and LSTM. Previously, we performed actor_loss.backward() and critic_loss.backward() separately. With LSTM modules, the history summarization module is shared by both actor and critic. The first time we call critic_loss.backward(), the computational graph that gets gradient information, including the LSTM module, is freed. This results in an error when we call actor_loss.backward(). This issue has been identified in here: https://github.com/facebookresearch/Pearl/issues/47

This diff fixes this issue by adding actor and critic losses and then apply backward to the sum.

Another fixed issue is that, when using OnPolicyEpisodicReplayBuffer, next_state is None, which results in an error in LSTM.

Another fixed issue is that, LSTM used the agent's current hidden state in learn_batch.

Finally, I changed LSTM to summarize a fixed-length history, instead of the entire history. Without this change, the algorithms can perform poorly.

Reviewed By: BillMatrix

Differential Revision: D52946247

fbshipit-source-id: 4c0a168deec33e25d57c2a9793ad22fb338fae6a