Move UCI data package away froms scripts
Summary:
This diff moves the file `utils/scripts/cb_benchmark/cb_download_benchmarks.py` to `utils`.

This was needed because the `scripts` directory is not included in the `pearl` library per se, so it was also not included in the `pearl` Bento kernel, therefore not being available on the Bento execution of the Contextual Bandits tutorial notebook that depends on that kernel.

Because the function to download UCI data can be thought of as a general utility and not part of a script, it's been moved to `util`. It has also been renamed to reflect the fact that it is not part of a benchmark script.

Reviewed By: Yonathae

Differential Revision: D54049997

fbshipit-source-id: 62403dfe156f246e246a29b774eb7de197d9dc8b